Fix DDS_HEADER and DDS_PIXELFORMAT
they used 64bit instead of 32bit fields, and since they are filled with a
brutal memcpy it didnt work.

Also sanitize some other parts of the code (added asserts())

git-svn-id: svn://svn.code.sf.net/p/therush/code@40 4b14b72a-3063-4cf0-bd6b-70bde676b5ec

+++ sources/librender/ddsOGL.h
@@ -8,6 +8,7 @@
 #include <string>
 #include <deque>
 #include <assert.h>
+#include <stdint.h>
 
 
 
@@ -73,30 +74,30 @@ namespace nv_dds
 
     struct DDS_PIXELFORMAT
     {
-        unsigned long dwSize;
-        unsigned long dwFlags;
-        unsigned long dwFourCC;
-        unsigned long dwRGBBitCount;
-        unsigned long dwRBitMask;
-        unsigned long dwGBitMask;
-        unsigned long dwBBitMask;
-        unsigned long dwABitMask;
+        uint32_t dwSize;
+        uint32_t dwFlags;
+        uint32_t dwFourCC;
+        uint32_t dwRGBBitCount;
+        uint32_t dwRBitMask;
+        uint32_t dwGBitMask;
+        uint32_t dwBBitMask;
+        uint32_t dwABitMask;
     };
 
     struct DDS_HEADER
     {
-        unsigned long dwSize;
-        unsigned long dwFlags;
-        unsigned long dwHeight;
-        unsigned long dwWidth;
-        unsigned long dwPitchOrLinearSize;
-        unsigned long dwDepth;
-        unsigned long dwMipMapCount;
-        unsigned long dwReserved1[11];
+        uint32_t dwSize;
+        uint32_t dwFlags;
+        uint32_t dwHeight;
+        uint32_t dwWidth;
+        uint32_t dwPitchOrLinearSize;
+        uint32_t dwDepth;
+        uint32_t dwMipMapCount;
+        uint32_t dwReserved1[11];
         DDS_PIXELFORMAT ddspf;
-        unsigned long dwCaps1;
-        unsigned long dwCaps2;
-        unsigned long dwReserved2[3];
+        uint32_t dwCaps1;
+        uint32_t dwCaps2;
+        uint32_t dwReserved2[3];
     };
 
     typedef enum TextureType
